% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_options.R
\name{all_options}
\alias{all_options}
\title{Denoting all compatible diplotype combinations per individual}
\usage{
all_options(dat, NEGs = NULL, POS.rep = TRUE)
}
\arguments{
\item{dat}{A vector. The genotyped data of the individuals, stating the whole genotype of a individual in one argument.}

\item{NEGs}{An optional character. How the `NEG` allele should be called, if \code{NULL} the `NEG` alleles will just be called `NEG`.}

\item{POS.rep}{A logical scalar. Whether or not all `POS` diplotype should be replaced by all other observed diplotypes. Individuals with \code{NA} as genotype will then also obtain all observed diplotypes.}
}
\value{
A list with each element stating all unique diplotypes of a individual that are compatible with the individual's genotype.
}
\description{
\code{all_options} is be used to transform genotype data of a single gene into a list containing all compatible diplotypes for that gene.
}
\details{
Different forms of ambiguity within the genotype data are recognized. A `+` indicates that there are multiple copies of a single gene in the individual's genotype; a `/` indicates allelic ambiguity, different allelic possibilities for the gene copy, and `|` indicates genotypic ambiguity, different genotypic possibilities for the complete genotype.
}
\examples{
dat = c("001/002+001/002", "003+004+005")
\dontrun{
all_options(dat = dat)
}

}
