% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.R
\name{B}
\alias{B}
\title{The second derivative of the likelihood for the Information Loss}
\usage{
B(HHOi, allele_freq)
}
\arguments{
\item{HHOi}{A vector. The values of an individual indicating how often the allele of interest occurs Homozygous-Heterozygous-Not.}

\item{allele_freq}{A numceric value. Allele frequencies of the allele of interest}
}
\value{
The outcome of the second derivative for the individual of the allele of interest.
}
\description{
Calculating the second derivative of the likelihood for the information loss. Support function of \code{\link{InformationLoss}}.
}
