% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing_AFs_in_HFs.R
\name{changing_AFs_in_HFs}
\alias{changing_AFs_in_HFs}
\title{Alter compound allele frequencies in haplotypes}
\usage{
changing_AFs_in_HFs(haplotypes, comb_vals, nonneg = TRUE)
}
\arguments{
\item{haplotypes}{A vector. Haplotypes and their probabilities.}

\item{comb_vals}{A vector. Compound haplotypes and their \emph{true} frequency}

\item{nonneg}{a logical scalar. Whether or not to make sure that there are no negative probabilities due to the resetting the compound haplotype frequency.}
}
\value{
A vector of the haplotypes and some altered probabilities.
}
\description{
For the profile EM-algorithm the compound allele frequency needs to be kept at bay. Support function for \code{\link{EM-algorithm}}.
}
