% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing_unresolvables.R
\name{changing_unresolvables}
\alias{changing_unresolvables}
\title{Combining alleles that cannot be distinguished from each other in the genotyping process}
\usage{
changing_unresolvables(dat, unresolvable)
}
\arguments{
\item{dat}{A data frame. The genotype calls of each individual as a separate row and each gene as a separate column.}

\item{unresolvable}{A three-column data frame. The first column must contain the corresponding gene; the second column the allele combination that cannot be distinguished and the third column the code to which that combination needs to be changed}
}
\value{
A vector with the unresolvable allele combinations combined into the indicated code.
}
\description{
To remove allelic ambiguities that originates from the fact that certain exons are not included in the sequencing process.
}
\details{
For consistency, if only a part of the unresolvable alleles is present in the individual's genotype call these alleles are still replaced by the indicated code.
}
