% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_frequency.R
\name{comb_frequency}
\alias{comb_frequency}
\title{Calculating the \emph{true} compound haplotype frequency}
\usage{
comb_frequency(
  EM_out = NULL,
  haplotypes = NULL,
  remain = NULL,
  combine = NULL,
  comb_adjust = FALSE
)
}
\arguments{
\item{EM_out}{An optional EM-algorithm analysis of \code{lst}, as provided by \code{\link{EM_algorithm}}. If \code{NULL}, than \code{haplotypes} must be specified.}

\item{haplotypes}{An optional vector. The haplotypes with their frequency, can be obtained by \code{\link{EM_algorithm}}. If \code{NULL}, than \code{EM_out} must be specified.}

\item{remain}{An optional vector. Haplotypes that are retained.}

\item{combine}{An optional vector. Haplotypes that are grouped.}

\item{comb_adjust}{A logical scalar. Whether or not the estimated \emph{true} compound haplotype frequency must be adjusted for the fact that the haplotype consists of multiple loci.}
}
\value{
A single value, representing the \emph{true} compound haplotype frequency.
}
\description{
Calculating the \emph{true} compound haplotype frequency based on a previous selection of which genes are grouped and which are retained.
}
\details{
Because of coding each \emph{x}-locus haplotype is required to have \emph{x} separate gene copies. Thus when a \emph{x}-locus hapltoype is combined into the compound haplotype, each gene copy is replaced by \emph{Comb1:x}. Since a haplotype frequency equals the product of its individual alleles. The compound haplotype frequency of a multi-locus haplotype needs to anticipate on that, requiring adjusting via \code{comb_adjust}.
}
