% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_AF.R
\name{combining_AF}
\alias{combining_AF}
\title{Grouping of the haplotypes based on their estimated frequencies}
\usage{
combining_AF(
  lst,
  EM_out = NULL,
  haplotypes = NULL,
  haplo_grouping = TRUE,
  comb_name = NULL,
  cumulative_threshold = NULL,
  cumulative_n = NULL,
  CO_thresh = 1e-07,
  CO_perc = 100,
  CO_min = 2,
  excluding_haplo = NULL,
  all_previous = TRUE,
  gene_subset = NULL
)
}
\arguments{
\item{lst}{A List. The data containing the possible diplotypes of each donor as a separate list element, can be obtained via \code{\link{all_options}}.}

\item{EM_out}{An optional EM-algorithm analysis of \code{lst}, as provided by \code{\link{EM_algorithm}}.}

\item{haplotypes}{An optional vector. The haplotypes with their frequency, can be obtained by \code{\link{EM_algorithm}}.}

\item{haplo_grouping}{A logical scalar. Whether or not grouping occurs on the haplotype level (\code{TRUE} is default) or on the allele level. In the case of 1 gene, haplotype and allele level are the same.}

\item{comb_name}{An optional character. Indicating the name of the combined group. If \code{NULL} the name is either `Comb` or `Comb` followed by a number (if previous combining step have been performed for \code{lst}).}

\item{cumulative_threshold}{An optional numeric value. The cumulative threshold value above which you want to group all alleles.}

\item{cumulative_n}{An optional integer. The number of haplotypes you want to retain, the \code{cumulative_n} haplotypes with the highest frequencies are retained.}

\item{CO_thresh}{An optional numeric value in \{0, 1\}. The cut-off threshold, only haplotypes lower than this threshold are be grouped.}

\item{CO_perc}{An optional numerical value in \{0, 100\}. The cut-off percentage, only the haplotypes belonging to the percentage of haplotypes with the lowest frequency are eligible for grouping.}

\item{CO_min}{An optional integer. The cut-off minimum, the minimum number of haplotypes which need to be grouped to actually group haplotypes.}

\item{excluding_haplo}{An optional vector. Which haplotypes never should be grouped, even if they do not meet the chosen criteria. NOTETOSELF: WAS excluding_geno, CAN BE WRONG IN OTHER FUNCTIONS}

\item{all_previous}{A logical scalar. Whether or not all previous compound haplotypes should be grouped in the newly formed compound haplotype. The name of all previous compound haplotypes should start with \emph{Comb}.}

\item{gene_subset}{An optional vector. Which genes need to be grouped if grouping_haplo = FALSE.}
}
\value{
A list with all possible genotypes of each donor, but now with some alleles replaced by `Comb`. Additionally are also the alleles which are combined into the `Comb` group provided.
}
\description{
Grouping of haplotypes into a common compound haplotype. There are multiple grouping criteria possible that can be chosen \enumerate{
  \item Based on a minimum frequency the haplotype should have (\code{CO_thresh}).
  \item Based on a number of haplotypes that should be kept (\code{cumulative_n}).
  \item Based on the cumulative frequency that should be kept (\code{cumulative_threshold}).
}. It is strongly advised to make use of the first criteria (\code{CO_thresh}) since this is for general use the most logical one. Further alterations are possible with other arguments.
}
\details{
With \code{CO_perc} and \code{CO_min} one can add extra limitations to the \code{CO_thresh} grouping criteria, but in practice their gain is limited and they primarily complicate the understanding of the grouping
}
