% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining_genes.R
\name{combining_genes}
\alias{combining_genes}
\title{Combining the diplotypes of two genes}
\usage{
combining_genes(x, y, haplo = TRUE, un_diplos = TRUE)
}
\arguments{
\item{x, y}{A list. The diplotype or haplotype list of a gene or of a combination of genes, can be obtained via \code{\link{all_options}}.}

\item{haplo}{a logical scalar. Whether or not \code{x} and \code{y} are combined on haplotype level, or are combined on genotype level (\code{TRUE} is default).}

\item{un_diplos}{a logical scalar. Whether or not only the unique diplotypes should be retained, or that each diplotype should occur the number of times it is created (\code{TRUE} is default).}
}
\value{
A list with the combined diplotypes of the two specified genes. The output in a non-ambiguous case is like this: GeneA1+GeneA2&GeneB1+GeneB2
}
\description{
To combine the information of two gene(s) combination(s), which is required for haplotype reconstruction based on multiple loci. Support function for \code{\link{HT_reconstruction}}.
}
\details{
If the genes are combined on haplotype level, the diplotypes notation is as follows: \emph{x1-y1+x2-y2} and \emph{x1-y2+x2-y1}. If the genes are combined on genotypic level, the notation is: \emph{x1+x2&y1+y2}.
}
