% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_vector.R
\name{cut_vector}
\alias{cut_vector}
\title{Split a numeric vector in subvectors}
\usage{
cut_vector(vect, vect_list = NULL)
}
\arguments{
\item{vect}{An integer vector. Must range from 1 till K.}

\item{vect_list}{An optional list. Each list element contains a integer (sub)vector to which a new integer needs to be added. Can be obtained via \code{cut_vector}.}
}
\value{
A list with each list element contains the K-1 subvectors.
}
\description{
To get all K-1 subvectors for a K vector. Support function for various functions.
}
