% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diplotype_contribution.R
\name{diplotype_contribution}
\alias{diplotype_contribution}
\title{Gives all diplotype configurations for a number of haplotypes}
\usage{
diplotype_contribution(nr_haplo)
}
\arguments{
\item{nr_haplo}{An integer. Number of unique haplotypes for the matrix.}
}
\value{
A matrix with all diplotypes as rows and all haplotypes as columns. The values indicate how often the haplotype contributes to the diplotype.
}
\description{
Give all theoretically possible diplotype configurations for a specified number of haplotypes. Support function for \code{\link{EM-algorithm}} and \code{\link{InformationLoss}}.
}
