% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dprime_to_Delta.R
\name{Dprime_to_Delta}
\alias{Dprime_to_Delta}
\title{Calculation of Delta_{ij}}
\usage{
Dprime_to_Delta(LD_out, p_i = NULL, q_j = NULL)
}
\arguments{
\item{LD_out}{The linkage disequilibrium matrix for the full K genes, as provided by \code{\link{LD}}.}

\item{p_i}{An optional vector. Haplotype frequencies for the K-1 genes.}

\item{q_j}{An optional vector. Haplotype frequencies for gene K.}
}
\value{
Vector with Delta_{ij} values.
}
\description{
Calculation of Delta_{ij}
}
