% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{EM_algorithm}
\alias{EM_algorithm}
\title{The Expectation-Maximization (EM-)algorithm}
\usage{
EM_algorithm(
  lst,
  nl = 200,
  freq_epsilon = 1e-05,
  likelihood_estimation = FALSE,
  likelihood_epsilon = 0,
  init_equal = TRUE,
  equal_first = FALSE,
  start_HTFs = NULL,
  regression = NULL,
  full_regression = FALSE,
  type_regression = "all",
  outcome = NULL,
  status = NULL,
  regression_reference = "NEG",
  comb_vals = NULL,
  noCNV = TRUE,
  nfolds = 50,
  alpha = 0.5,
  lambda = NULL,
  split_evenly = TRUE,
  weighted_predictor = TRUE,
  CNV_option = "second",
  probYH_epsilon = 1e-10
)
}
\arguments{
\item{lst}{A list. Data containing the possible diplotypes of each donor as a separate list element, can be obtained via \code{\link{all_options}}.}

\item{nl}{An integer. Total number of iterations till the EM-algorithm is terminated, default = 200.}

\item{freq_epsilon}{An numeric value. The minimal difference the biggest changing haplotype frequencies between two subsequent iterations must have to keep the EM-algorithm going. If the biggest difference is smaller, the EM-algorithm will be terminated.}

\item{likelihood_estimation}{A logical scalar. Whether or not the likelihood is calculated each iteration (\code{FALSE} is default). This is quite computational intensive.}

\item{likelihood_epsilon}{A numeric value. The minimal difference the changing likelihood between two subsequent iterations must have to keep the EM-algorithm going. Only relevant if \code{likelihood_estimation} is \code{TRUE}.}

\item{init_equal}{A logical scalar. Whether or not the probabilities in the initial value matrix are equal (a 0 when diplotype is not compatible with individuals genotype call, and a 1 when the diplotype is compatible) or that the number of times a compatible diplotype is stated in \code{lst} equals the value the initial value matrix should have (\code{TRUE} is default).}

\item{equal_first}{A logical scalar. Whether or not the estimated haplotype frequencies of the first iteration are equal (\code{FALSE} is default).}

\item{start_HTFs}{An optional vector. Haplotype frequencies where the diplotype frequencies of the first iteration are started with. If a haplotype is not specified here, it will get a frequency equal to half of the minimum supplied frequency}

\item{regression}{An optional character. Multiple options are possible: \code{NULL} (default), \code{linear}, \code{lin_ridge}, \code{logistic}. If not \code{NULL} forces the EM-algorithm to run a regression analysis each iteration which is used for haplotype frequency estimation. Then requires an outcome for each individual in \code{outcome}.  If \code{NULL}, haplotype frequency estimation will proceed without the regression analyses.}

\item{full_regression}{A logical scalar. Whether or not the regression models should contain all haplotype predictors next to the allelic predictors. Only functional when \code{regression} is not \code{NULL}. NOTETOSELF: Dit is nu waarschijnlijk al overbodig.}

\item{type_regression}{An optional character. If the EM-algorithm with outcome is requested, which inside EM-algorithm regression model needs to be run (either "alleles", "backward", "penalized"). One an also run all three by supplying "all".}

\item{outcome}{An optional vector. The outcome of each individual present in \code{lst}. Only functional when \code{regression} is not \code{NULL}.}

\item{status}{An optional vector. The status of each individual present in \code{lst}, only applicable for survival analysis. Only functional when \code{regression} is not \code{NULL}.}

\item{regression_reference}{A character. Which predictor, or which predictors, should be used as reference group in the regression models.}

\item{comb_vals}{An optional vector. The \emph{true} frequency of the combinational haplotypes is, thus at which values the \code{comb_vals} should be fixed in each iteration.}

\item{noCNV}{A logical scalar. Whether or not the copy number variation haplotypes should be accomodated in the allelic predictors or should be kept as separate predictors. Only functional when \code{regression} is not \code{NULL}.}

\item{nfolds}{An optional integer. The group size of the cross-validation in \code{\link[glmnet]{cv.glmnet}} used for the penalized regression models. Only functional when \code{regression} is not \code{NULL}.}

\item{alpha}{An optional numeric value in \{0, 1\}. Which penalized regression should be performed. If \code{alpha} is 0 the ridge penalty is used, when \code{alpha} is 1 the LASSO penalty is applied and when \code{alpha} (0, 1) the elastic-net is used. Only functional when \code{regression} is not \code{NULL}.}

\item{lambda}{An optional integer. The number of standard errors that should be added to the optimal lambda, chosen by \code{\link[glmnet]{cv.glmnet}}. Literature suggest that lambda is 1 is the best choice, if \code{NULL}, this will be chosen. Only functional when \code{regression} is not \code{NULL}.}

\item{split_evenly}{An optional logical scalar. Whether or not for the calculation of the probability of the outcome given the diplotype (probYH) the probabilities are equally divided over the combinational and the non-combinatonal haplotypes, or that the combinational haplotype gets its \emph{true} frequency. Note that this \emph{true} frequency is only the true frequency for the frequency, not for probYH.}

\item{weighted_predictor}{A logical scalar. Whether or not each observed diplotype of an individual should get its own row where its frequency is added as a weight in the regression models (\code{TRUE} is default).}

\item{CNV_option}{A character. Indicating what to do with the copy number variation alleles with the inclusion of the outcome (either "first", "second", or "third").}

\item{probYH_epsilon}{A numeric value. Threshold value below which it is assumed that the outcome information is purely noise and should be discarded for a given individual.}
}
\value{
Returns both the fully updated probability matrix with the probability vector for each donor, as the calculated haplotype frequencies for each iteration for each haplotype
}
\description{
The EM-algorithm that underlies the complete profile EM-algorithm from \code{\link{HT_reconstruction}}. Haplotype frequencies are estimated in each iteration until convergence. Haplotype frequencies can be estimated
}
