% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluding_genotypes.R
\name{excluding_genotypes}
\alias{excluding_genotypes}
\title{Excluding specific diplotypes}
\usage{
excluding_genotypes(
  lst,
  remove_alleles = NULL,
  remove_genotypes = NULL,
  no_hat3plus = FALSE
)
}
\arguments{
\item{lst}{A list. All possible diplotypes of a donor, as obtained via \code{\link{all_options}}}

\item{remove_alleles}{An optional vector. If a diplotype contains any of these allele(s), they will be discarded.}

\item{remove_genotypes}{An optional vector. Which diplotypes need to be discarded.}

\item{no_hat3plus}{A logical scalar. Whether or not all diplotypes that have a haplotype that consists of three or more genes need to be discarded (\code{FALSE} is default).}
}
\value{
A list with the remaining possible diplotypes of a donor. There is also a vector with all removed diplotypes and a vector with the donors that have no diplotypes left.
}
\description{
To exclude specific diplotypes from the diplotype lists of individuals. Either by fully specifying the diplotypes to discard, or by specifying a allele that, if present in a diplotype, will discard all diplotypes.
}
