% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_information_haplotypes.R
\name{external_information_haplotypes}
\alias{external_information_haplotypes}
\title{Selecting the relevant haplotype frequencies from an external dataset}
\usage{
external_information_haplotypes(ext_dat, genes)
}
\arguments{
\item{ext_dat}{A matrix. Allelic content of each gene in a separate column and the frequencies in a separate column.}

\item{genes}{A vector. Genes for which the haplotype frequency needs to be determined.}
}
\value{
A matrix with two columns, the first column denotes the unique haplotypes and the second column the corresponding frequencies.
}
\description{
To calculate the haplotype frequencies for only a subset of genes. By discaring irrelevant genes, dubble haplotypes will occur and need to aggregate over them.
}
