% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_order.R
\name{grouping_order}
\alias{grouping_order}
\title{Determines which gene is the best choice to add into the profile EM-algorithm reconstruction}
\usage{
grouping_order(freqs1, freqs2, comb_group = NULL, freqs1_no_comb = FALSE)
}
\arguments{
\item{freqs1}{A vector. Allele frequencies of the analysis of the starting gene without other genes.}

\item{freqs2}{A vector. The allele frequencies of the starting gene in a grouping configuration.}

\item{comb_group}{An optional vector. Containing the names of the alleles grouped in the compound haplotype. If not \code{NULL} the frequency of the compound haplotype is equally distributed over all alleles.}

\item{freqs1_no_comb}{A logical scalar. Whether or not the compound haplotype frequency of \code{freqs2} is equally distributed over all alleles that are not grouped in \code{freqs1}, but are grouped in \code{freqs2} (\code{FALSE} is default).}
}
\value{
A list with the mean difference between the alleles and the names of the alleles which are used for calculating the difference.
}
\description{
Estimates the mean squared difference between the allele frequencies of the starting genes with and without the addition of the the new genes. The gene that is added into the reconstruction maximizes the mean squared difference. Support function of \code{grouping_order_all}.
}
