% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_order_all.R
\name{grouping_order_all}
\alias{grouping_order_all}
\title{Profile EM-algorithm for a single starting gene}
\usage{
grouping_order_all(
  gene1,
  ...,
  other_genes = NULL,
  gene_names = NULL,
  orderings = NULL,
  comb_vals = NULL,
  naive = TRUE,
  fixed_order = FALSE,
  candidate_list = NULL,
  comb_gene1 = TRUE,
  combine_haplos = TRUE,
  CO_thresh = 1e-07,
  CO_perc = 100,
  CO_min = 1,
  all_previous = TRUE,
  outcome_EM = FALSE,
  regression = "gaussian",
  full_regression = FALSE,
  type_regression = "all",
  regression_reference = "NEG",
  outcome = NULL,
  status = NULL,
  noCNV = TRUE,
  CNV_option = "second",
  nfolds = 50,
  alpha = 0.5,
  lambda = NULL,
  weighted_predictor = TRUE
)
}
\arguments{
\item{gene1}{A list. Diplotype list of the starting gene, can be obtained via \cofull_regressionde{\link{all_options}}.}

\item{...}{The diplotype lists of the other genes can also be supplied without specification of \code{other_genes}.}

\item{other_genes}{An optional list. Diplotype list of the other genes that need to be added into the reconstruction, can be obtained via \code{\link{all_options}}.}

\item{gene_names}{An optional vector. Names of genes, need to be in the order in which they are supplied.}

\item{orderings}{An optional vector. Numerical order of the genes.}

\item{comb_vals}{An optional vector. The \emph{true} frequency of the combinational haplotypes is, thus at which values the \code{comb_vals} should be fixed in each iteration.}

\item{naive}{A logical scalar. Whether or not the naive analyses also should be run with the data}

\item{fixed_order}{A logical scalar. Whether or not the order in which the genes are grouped is based on the order in which they are supplied, or that the order is based on some heuristic strategy.}

\item{candidate_list}{An optional vector. Which haplotypes are never allowed to be grouped into the compound haplotype.}

\item{comb_gene1}{A logical scalar. Whether or not the alleles of the starting gene need to be grouped before the reconstruction starts (\code{TRUE} is default).}

\item{combine_haplos}{A logical scalar. Whether or not low-frequency haplotypes in the haplotype reconstruction are grouped (\code{TRUE} is default).}

\item{CO_thresh}{An optional numeric value \in \{0, 1\}. The cut-off threshold, only haplotypes lower than this threshold are be grouped.}

\item{CO_perc}{An optional numerical value \in \{0, 100\}. The cut-off percentage, only the haplotypes belonging to the percentage of haplotypes with the lowest frequency are eligible for grouping.}

\item{CO_min}{An optional integer. The cut-off minimum, the minimum number of haplotypes which need to be grouped to actually group haplotypes.}

\item{all_previous}{A logical scalar. Whether or not all previous compound haplotypes should be grouped in the newly formed compound haplotype. The name of all previous compound haplotypes should start with \emph{Comb}.}

\item{regression}{An optional character. Multiple options are possible: \code{NULL} (default), \code{linear}, \code{lin_ridge}, \code{logistic}. If not \code{NULL} forces the EM-algorithm to run a regression analysis each iteration which is used for haplotype frequency estimation. Then requires an outcome for each individual in \code{outcome}.  If \code{NULL}, haplotype frequency estimation will proceed without the regression analyses.}

\item{full_regression}{A logical scalar. Whether or not the regression models should contain all haplotype predictors next to the allelic predictors. Only functional when \code{regression} is not \code{NULL}. \lars{Dit is nu waarschijnlijk al overbodig.}}

\item{regression_reference}{A character. Which predictor, or which predictors, should be used as reference group in the regression models.}

\item{outcome}{An optional vector. The outcome of each individual present in \code{lst}. Only functional when \code{regression} is not \code{NULL}.}

\item{status}{An optional vector. The status of each individual present in \code{lst}, only applicable for survival analysis. Only functional when \code{regression} is not \code{NULL}.}

\item{noCNV}{a logical scalar. Whether or not the copy number variation haplotypes should be accomodated in the allelic predictors or should be kept as separate predictors. Only functional when \code{regression} is not \code{NULL}.}

\item{nfolds}{An optional integer. The group size of the cross-validation in \code{\link[glmnet]{cv.glmnet}} used for the penalized regression models. Only functional when \code{regression} is not \code{NULL}.}

\item{alpha}{An optional numeric value \in \{0, 1\}. Which penalized regression should be performed. If \code{alpha} is 0 the ridge penalty is used, when \code{alpha} is 1 the LASSO penalty is applied and when \code{alpha} (0, 1) the elastic-net is used. Only functional when \code{regression} is not \code{NULL}.}

\item{lambda}{An optional integer. The number of standard errors that should be added to the optimal lambda, chosen by \code{\link[glmnet]{cv.glmnet}}. Literature suggest that \lambda is 1 is the best choice, if \code{NULL}, this will be chosen. Only functional when \code{regression} is not \code{NULL}.}
}
\value{
While running the different combinations and iterations, the order in which the genes are combined are concatenated. After each inclusion, the included gene is removed from the list and cannot be included again. The output is all grouping information of the last grouping and the retained haplotypes, additionally, EM-algorithm output for the intermediate and last reconstruction is supplied.
}
\description{
The best order in which the genes are added into the haplotype reconstruction is based via some heuristic strategy. The gene in \code{gene1} is the starting gene, and all other genes that need to be added into the reconstruction are denoted in \code{other_genes}. The profile EM-algorithm is described in \cite{}.
}
