% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_order_iter.R
\name{grouping_order_iter}
\alias{grouping_order_iter}
\title{Single reconstruction step for the profile EM-algorithm.}
\usage{
grouping_order_iter(
  gene1,
  ...,
  reference_info = NULL,
  gene_names = NULL,
  orderings = NULL,
  comb_vals = NULL,
  all_previous = FALSE,
  fixed_order = FALSE,
  candidate_list = NULL,
  combine_haplos = TRUE,
  CO_thresh = 1e-07,
  CO_perc = 100,
  CO_min = 2
)
}
\arguments{
\item{gene1}{A list. Diplotype list of the starting gene, can be obtained via \code{\link{all_options}}.}

\item{...}{The diplotype lists of the other genes can also be supplied without specification of \code{other_genes}.}

\item{reference_info}{An optional list. Contains the frequencies and diplotypes in the combined group, required for determining the grouping order. If no grouping is required for the reference group, the `comb` needs to be specified as NULL. If no vector is supplied, the analysis of gene1 alone will be used.}

\item{gene_names}{An optional vector. Names of genes, need to be in the order in which they are supplied.}

\item{orderings}{An optional vector. Numerical order of the genes.}

\item{comb_vals}{An optional vector. The \emph{true} frequency of the combinational haplotypes is, thus at which values the \code{comb_vals} should be fixed in each iteration.}

\item{all_previous}{A logical scalar. Whether or not all previous compound haplotypes should be grouped in the newly formed compound haplotype. The name of all previous compound haplotypes should start with \emph{Comb}.}

\item{fixed_order}{A logical scalar. Whether or not the order in which the genes are grouped is based on the order in which they are supplied, or that the order is based on some heuristic strategy.}

\item{candidate_list}{An optional vector. Which haplotypes are never allowed to be grouped into the compound haplotype.}

\item{combine_haplos}{A logical scalar. Whether or not low-frequency haplotypes in the haplotype reconstruction are grouped (\code{TRUE} is default).}

\item{CO_thresh}{An optional numeric value in \{0, 1\}. The cut-off threshold, only haplotypes lower than this threshold are be grouped.}

\item{CO_perc}{An optional numerical value in \{0, 100\}. The cut-off percentage, only the haplotypes belonging to the percentage of haplotypes with the lowest frequency are eligible for grouping.}

\item{CO_min}{An optional integer. The cut-off minimum, the minimum number of haplotypes which need to be grouped to actually group haplotypes.}
}
\value{
The output of \code{\link{grouping_order}} for each of the grouping configurations with the starting gene and the gene added into the reconstruction (with analysis).
}
\description{
Determines which gene is best suitable to be added into the reconstruction next. Analyses are run with this gene added into the reconstruction. Support function of \code{\link{grouping_order_all}},
}
