% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplo2AF.R
\name{Haplo2AF}
\alias{Haplo2AF}
\title{Calculate allele frequencies from haplotype frequencies}
\usage{
Haplo2AF(EM_out = NULL, haplotypes = NULL)
}
\arguments{
\item{EM_out}{An optional EM-algorithm analysis of \code{lst}, as provided by \code{\link{EM_algorithm}}.}

\item{haplotypes}{An optional vector. Containing the haplotype frequencies.}
}
\value{
An list with in each list element the allele frequencies of the different genes. The first element contains the AFs of the first gene, the second element of the second gene etc.
}
\description{
To calculate allele frequencies for all genes based on the haplotype frequencies. Either an EM-algorithm analysis or the haplotype frequencies need to be supplied.
}
