% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HF_2_DF.R
\name{HF_2_DF}
\alias{HF_2_DF}
\title{Calculate frequencies for all diplotypes}
\usage{
HF_2_DF(Diplotypes, haplo_freq)
}
\arguments{
\item{Diplotypes}{A vector. Diplotype names for which the frequency needs to be estimated.}

\item{haplo_freq}{A vector. Haplotypes and their frequency.}
}
\value{
A vector with diplotype frequencies.
}
\description{
Calculation of diplotype frequencies based on haplotype frequencies.
}
