% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_reconstruction_criteria.R
\name{HT_reconstruction_criteria}
\alias{HT_reconstruction_criteria}
\title{Individual haplotype reconstruction measure}
\usage{
HT_reconstruction_criteria(
  EM_mat = NULL,
  naive_mat = NULL,
  correct_mat,
  measure1 = TRUE,
  noCNV = TRUE
)
}
\arguments{
\item{EM_mat, naive_mat}{An optional matrix. Denoting for each individual the probabilities estimated for each possible diplotypes, as can be obtained via \code{EM-algorithm} or \code{Naive_analysis}. One of the two must be specified.}

\item{correct_mat}{A matrix. The matrix with the correct probabilities for each diplotype, thus each individual has one 1 for the actual simulated diplotype and a 0 for all other diplotypes.}

\item{measure1}{A logical scalar, whether or not we are interested only in the completely correct DT or that we are interested in all diplotypes and consider the diplotypes that have alleles that are partly correct to be correct according to ratio (\code{TRUE} is default).}

\item{noCNV}{A logical scalar, whether or not we want to split the CNVs into separate alleles (\code{TRUE} is default).}
}
\value{
The sum of squares measure of the chosen haplotype reconstruction measure.
}
\description{
Multiple estimates of the individual haplotype reconstruction measure, where the individual's probability vector is compared to the truth. This haplotype reconstruction measure gives a better indication on how well the haplotype reconstruction methods perform on an individual level.
}
\details{
There are different haplotype reconstruction measures possible, depending on which arguments are specified. \itemize{
\item Compare the probability estimated for the correct diplotype with 1. Executed when \code{only_corr} is \code{TRUE}.
\item Compare the probabilities estimated for all compatible diplotypes, diplotypes that have alleles that are partly correct are considered to be correct according to ratio. Executed when \code{only_corr} is \code{FALSE} and \code{DT_error} is \code{TRUE}.
\item Compare the probabilities estimated for all compatible diplotypes, haplotypes that have alleles that are partly correct are considered to be correct according to ratio. Executed when \code{only_corr} is \code{FALSE} and \code{DT_error} is \code{FALSE}.
}
  The second option is considered to be the best measure to summarize the individual haplotype frequencies when no haplotype effects are considered. Diplotypes consisting of different haplotype configurations (which the algorithm cannot distinguish between) are not punished this way.
}
