% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{information_loss}
\alias{information_loss}
\title{Calculate the Information loss for a gene}
\usage{
information_loss(
  EM_out = NULL,
  HTFs = NULL,
  EM_mat = NULL,
  haplotype_level = FALSE,
  nr_gene = 1
)
}
\arguments{
\item{EM_out}{An optional EM-algorithm analysis of \code{lst}, as provided by \code{\link{EM_algorithm}}.}

\item{HTFs}{An optional vector. Haplotype frequencies for all haplotypes, as are estimated in \code{\link{EM_algorithm}}. \lars{was haplo_freqs}}

\item{EM_mat}{An optional matrix. Probabilities for each compatible diplotype in the dataset, as can be provided by \code{\link{EM_algorithm}}.}

\item{nr_gene}{An integer. The number of genes that are in the reconstruction.}
}
\value{
An matrix with the calculated I_{Y}, I_{X}, ratio I_{Y} / I_{X} and the effective sample size (I_{Y} $\theta (1 - \theta)$) for the specified gene
}
\description{
Calculates the amount of information loss within a reconstruction, as discussed in \cite{}.
}
