% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_modeling.R
\name{iterative_modeling}
\alias{iterative_modeling}
\title{Running regression models with only significant haplotypes}
\usage{
iterative_modeling(
  outcome,
  EM_mat = NULL,
  weighted_mat = NULL,
  regression = "gaussian",
  regression_reference = "NEG",
  only_sign_EES = TRUE,
  pval_thresh = 0.1,
  candidate_list = NULL,
  noCNV = TRUE,
  CNV_option = "second",
  gene_indication = FALSE
)
}
\arguments{
\item{EM_mat}{An optional matrix. Denoting for each individual the probabilities estimated for each possible diplotypes, as can be obtained via \code{EM-algorithm}.}

\item{regression}{An optional character. Multiple options are possible: \code{NULL} (default), \code{gaussian} and \code{logistic}. If not \code{NULL} forces the EM-algorithm to run a regression analysis each iteration which is used for haplotype frequency estimation. Then requires an outcome for each individual in \code{outcome}.  If \code{NULL}, haplotype frequency estimation will proceed without the regression analyses.}

\item{regression_reference}{A character. Which predictor, or which predictors, should be used as reference group in the regression models.}

\item{only_sign_EES}{A logical scalar. Whether or not multi-locus haplotypes are only allowed to be constructed from significant alleles.}

\item{pval_thresh}{A numeric value \in \{0, 1\}. Multi-locus haplotypes with p-values below this threshold value are retained in the regression model as predictors.}

\item{candidate_list}{A list. Overview of haplotypes that need to be predictors in the regression models.}

\item{noCNV}{a logical scalar. Whether or not the copy number variation haplotypes should be accomodated in the allelic predictors or should be kept as separate predictors. Only functional when \code{regression} is not \code{NULL}.}

\item{yi}{An optional vector. The outcome of each individual present in \code{lst}. Only functional when \code{regression} is not \code{NULL}.}
}
\value{
A list where each list element contains the estimated betas from each regression step.
}
\description{
Regression models with only the allele effects are appropriate when the effects are only attributed to alleles, but additional effects that occur when specific alleles come together in a haplotype. Therefore certain haplotypes will need to be included as predictors in the regression models. Because of the abundance of haplotypes, only a limited number of haplotypes are allowed. Via backward selection the haplotypes with the lowest p-values are retained. Multi-loci haplotypes are only considered if one of its subset haplotypes are retained in the predictor matrix.
}
