% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_reconstruction.R
\name{KIR_analysis}
\alias{KIR_analysis}
\title{Haplotype reconstruction via the profile EM-algorithm}
\usage{
KIR_analysis(
  gene_GTs = NULL,
  gene_lsts = NULL,
  nr_start_genes = NULL,
  gene_order = NULL,
  CO_thresh = 1e-07,
  all_previous = TRUE
)
}
\arguments{
\item{gene_GTs}{An optional data frame. Each column contains the genotype calls of a separate gene. If \code{NULL} then \code{gene_lsts} must be supplied.}

\item{gene_lsts}{A optional list. Each list element has the compatible DTs, as can be provided by \code{\link{all_options}}. If \code{NULL} then \code{gene_GTs} must be supplied, if both are supplied, \code{gene_lsts} is overwritten by input of \code{gene_GTs}.}

\item{nr_start_genes}{An optional integer. The number of genes that should be used as a starting gene. The reconstruction is executed with the indicated number of starting genes denoted in the order of \link{gene_lsts}. As default each gene will be used as starting gene once.}

\item{gene_order}{A optional vector indicating the order in which the results of each reconstruction should be reordered.}

\item{CO_thresh}{A numerical value. Threshold value that determines which alleles or haplotypes are rare and will be grouped in the compound HT (\emph{$1e^{-07}$} is default).}

\item{all_previous}{A logical scalar. Whether or not all old combinational haplotypes should be forced to be grouped in the new combinational haplotype in a new reconstruction step if their probability is higher than \link{CO_thresh} (\code{TRUE} is default).}
}
\value{
\code{HT_reconstruction} returns a list with two elements: 
\describe{
\item{Matrices}{a matrix containing the estimated diplotype frequencies for each donor. The diplotypes of each starting gene are reordered into the common gene order}
\item{Frequencies}{the haplotype frequencies estimated for the final reconstructions. The haplotypes of each starting gene are reordered into the common gene order}
}
}
\description{
General function to run the profile EM-algorithm as described in \cite{}.
}
\examples{
gene1 = list(c("A+A", "A+B", "B+B"), "C+D", c("A+D", "C+D"))
gene2 = list(c("E+E", "E+F"), c("E+E", "E+F", "F+F"), "E+F")
gene3 = list("G+H", "G+H", c("G+G", "G+H", "H+H"))

gene_lsts = list("Gene1" = gene1, "Gene2" = gene2, "Gene3" = gene3)

\dontrun{
HT_reconstruction(gene_lsts, nr_start_genes = 2, gene_order = names(gene_lsts))
}

}
\references{

}
