% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_divergence.R
\name{KL_divergence}
\alias{KL_divergence}
\title{Kullback-Leibler divergence measure}
\usage{
KL_divergence(P, Q, comb_add = FALSE, base = 2)
}
\arguments{
\item{P}{A vector. Values to compare with \code{Q}.}

\item{Q}{A vector. Correct values, need to consists of the same value names as \code{P}.}

\item{comb_add}{A logical scalar. Whether or not the cumulative sum of the vectors \code{P} and \code{Q} need to be composed of a compound haplotype, or that both vectors need to be normalized.}

\item{base}{An integer. Base value for the logaritmic transformation in \code{\link[philentropy]{KL}}.}
}
\value{
The kullback-leibler divergence measure between \code{P} and \code{Q}.
}
\description{
Calculation of the Kullback-Leibler divergence measure between two vectors. Two vectors \code{P} and \code{Q} will be standardized, so not necessarily need to sum to 1. Actual KLD measure is executed by \code{\link[philentropy]{KL}}.
}
