% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_frame.R
\name{make_data_frame}
\alias{make_data_frame}
\title{Makes a data frame from a list}
\usage{
make_data_frame(lst)
}
\arguments{
\item{lst}{A list. The genotype options splitted (\emph{e.g.} via \code{\link[base]{strsplit}}), on the allelic ambiguities (`/`).}
}
\value{
A data_frame with the options of each gene copy as a different row. The number of columns are equal to the gene copy with the most options, other rows are supplemented with \code{NA}.
}
\description{
Support function for \code{\link{Splitting_slashes}}.
}
