% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_gene_copies.R
\name{multiple_gene_copies}
\alias{multiple_gene_copies}
\title{Handling all possible gene copy genotype of a genotype with at least one "+", including the multi-allele haplotypes}
\usage{
multiple_gene_copies(xx, NEGs = NULL)
}
\arguments{
\item{xx}{A character. The possible genotype option of the individual.}
}
\value{
All possible genotypes with the multi-allele haplotype
}
\description{
Support function of \code{\link{all_options}}. Function can only correctly handle an argument that has a `+` and cannot correctly handle arguments that contain allelic (`/`) or genotypic ambiguities (`|`).
}
