% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_single_options.R
\name{multiple_single_options}
\alias{multiple_single_options}
\title{Handling the genotype results for multiple single gene copy genotypes}
\usage{
multiple_single_options(xx, sign, NEGs = "NEG")
}
\arguments{
\item{xx}{A character. The possible genotype option of the individual. The genotype of the i-th donor, can not handle "+".}

\item{sign}{A character. On which sigh the genotype must be split, can be either "/" (allelic ambiguities) or "|" (genotypic ambiguities).}

\item{NEGs}{An optional character. How the `NEG` allele should be called, by default it will be called `NEG`.}
}
\value{
All possible genotype for the single gene copy genotype.
}
\description{
Support function of \code{\link{all_options}}. Function can only correctly handle single gene copies with allelic (`/`) or genotypic ambiguities (`|`) and cannot correctly handle arguments with a `+`.
}
