% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naive_analysis.R
\name{Naive_analysis}
\alias{Naive_analysis}
\title{Naive HT reconstruction analysis}
\usage{
Naive_analysis(lst, haplo = TRUE, CO_thresh = 1e-07)
}
\arguments{
\item{lst}{A list. Data containing the possible diplotypes of each donor as a separate list element, can be obtained via \code{\link{all_options}}.}

\item{haplo}{A logical scalar. Whether or not diplotype frequencies are estimated by first estimating haplotype frequencies, or immediately start by estimating diplotype frequencies (\code{TRUE} is default).}

\item{CO_thresh}{A numeric value in \{0, 1\}. The cut-off threshold, only haplotypes lower than this threshold are be collapsed.}
}
\value{
A choice vector for the dictatorial naive analysis and a probability matrix for the democratic naive analysis. For both methods also haplotype frequencies are returned if \code{haplo} is \code{TRUE}
}
\description{
Runs two naive analyses to be compared to the analysis with the EM-algorithm based \code{\link{HT_reconstruction}}.
}
