% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nested_loop_plot.R
\name{Nested_loop_plot}
\alias{Nested_loop_plot}
\title{Nested-loop plot}
\usage{
Nested_loop_plot(
  Scenario_overview,
  EM_betas,
  naive_betas = NULL,
  list_variable = "outcome",
  HTR_name = NULL,
  variables = c("naive", "LD", "Amb"),
  mean_plot = FALSE,
  mean_name = "mean",
  title = NULL,
  yname = "RMSE",
  outlier_prob = 0.99,
  maxY = NULL,
  extra_lines = NULL,
  second_Yaxis = NULL,
  altered_replications = NULL
)
}
\arguments{
\item{Scenario_overview}{A matrix. Overview of the characteristics of all simulation scenarios}

\item{EM_betas}{A list. Each list element are the values to plot obtained via a different EM-algorithm analysis method (\emph{e.g.,} with vs without outcome analysis).}

\item{naive_betas}{An optional list. Each list element are the values to plot obtained via a different naive analysis method (\emph{e.g.,} dictorial vs democratic analysis).}

\item{list_variable}{An optional character. Variable name that explains the different list elements of \code{EM_betas} and \code{naive_betas}.}

\item{variables}{A vector. The variables where the simulation scenarios differ from each other. Their values must be stated in \cite{Scenario_overview}.}

\item{mean_plot}{A logical scalar. Whether or not the mean value of the values to plot for each scenario must be plotted as a horizontal line.}

\item{title}{An optional vector. Title of the Nested-loop plot.}

\item{yname}{A character. The name of the y-axis, how the values to plot are calculated.}

\item{outlier_prob}{An optional numeric value \in \{0, 1\}. Indicating which part of the distribution is discarded and placed as empty triangles on top of the graph.}

\item{maxY}{An optional numeric value. A crude cut-off value, replacing the one calculated with \code{outlier_prob}. Less sophisticated and not data-driven.}

\item{extra_lines}{An optional list. Additional values to plot as colored horizontal lines for each scenario. Predominantly used for the individual haplotype frequencies.}

\item{second_Yaxis}{An optional vector. If supplied, the values of the y-axis will be transformed with the specified value and also plotted on the right. Useful if the values in \code{extra_lines} have been scaled. The name of \code{second_Yaxis} will be used as the name for the second y-axis.}

\item{altered_replications}{An optional vector. Indicates per scenario how many replications have been replaced due to the Euclidean distance estimates.}
}
\value{
The nested-loop plot and the order in which the scenarios are plotted
}
\description{
Creating of a nested-loop plot to visualize the estimated haplotype frequencies and effect sizes for a large number of simulation scenarios. The nested-loop plot is based on \cite{}.
}
\details{
... For the cut-off value of the outliers, \code{outlier_prob} is used as quantile for the distribution of all values to plot in \code{EM_betas} + three times the interquantile range. All values above this cut-off value are replaced by empty triangles and plotted with the cut-off value as their \emph{true} value. Another option would be to manually choose the cut-off value via \code{maxY}. But this is less elegant, because then the value is not based on the data and
}
