% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_coding.R
\name{occurrence_coding}
\alias{occurrence_coding}
\title{Making an occurrence matrix of haplotype occurrences}
\usage{
occurrence_coding(
  lst = NULL,
  EM_out = NULL,
  EM_out_mat = NULL,
  haplos_keep = NULL,
  gene_indication = TRUE,
  noCNV = FALSE,
  CNV_option = "second"
)
}
\arguments{
\item{lst}{An optional list with each individual having a separate list elements with all its compatible diplotypes, as provided by \code{\link{all_options}}}

\item{EM_out}{An optional list of the EM-algorithm output of the \code{lst} compatible diplotype list}

\item{EM_out_mat}{The optional matrix of the EM-algorithm of the \code{lst} compatible diplotype list}

\item{haplos_keep}{An optional vector. For which haplotypes the occurrence matrices need to be made. if \code{NULL} the occurrence matrix will be based on all supplied haplotypes}

\item{gene_indication}{A logical scalar. Whether or not it is indicated from which gene combination the allele or haplotype is (default is \emph{TRUE}).}

\item{noCNV}{a logical scalar. Whether or not the copy number variation haplotypes should be accommodated in the allelic predictors or should be kept as separate predictors. Only functional when \code{regression} is not \code{NULL}.}

\item{CNV_option}{A character. Indicating what to do with the copy number variation alleles with the inclusion of the outcome (either "first", "second", or "third").}
}
\value{
A matrix with each individual having a separate row and each haplotype a separate column. In a full-information setting, a 2 is assigned if the haplotype occurs two times in the individual's diplotype, a 1 if it occurs once and a 0 if it does not occur. With ambiguity, values between 0 and 2 can occur, while the rowsum is always equal to 2.
}
\description{
To make an occurrence matrix indicating how often each haplotype occurs in the individual's diplotypes.
}
