% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_significance.R
\name{penalized_significance}
\alias{penalized_significance}
\title{P-value calculation for penalized regression}
\usage{
penalized_significance(pred_mat, betas, lambda, outcome)
}
\arguments{
\item{pred_mat}{A matrix. Predictor matrix used in the penalized regression model.}

\item{betas}{A numeric value. Betas estimated in the penalized regression model with the chosen \code{lambda}.}

\item{lambda}{A numeric value. Lambda chosen for the penalized regression model.}

\item{outcome}{A vector. The outcome of each individual present in \code{lst}. Only functional when \code{regression} is not \code{NULL}.}
}
\value{
A vector with estimated p-values for the predictors.
}
\description{
Due to the bias introduced in the estimates due to the penalty, p-values are not (entirely) valid. However, they can be estimated, as described in \cite{Cule et al., 2011}. Support function for \code{\link{iterative_modeling}}.
}
