% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducing_haplo.R
\name{reducing_haplo}
\alias{reducing_haplo}
\title{Reducing haplotype frequencies into the K-1 genes and gene K}
\usage{
reducing_haplo(haplo_freqs)
}
\arguments{
\item{haplo_freqs}{A vector. The frequencies of the K-gene haplotypes.}
}
\value{
The haplotype frequencies of the K-1 gene combinations and the frequencies of gene K, also the total number of genes (K) and the genes in the combination (K-1) are denoted.
}
\description{
The haplotype frequencies will be split in two groups, the K-1 genes and gene K. Supporting function for \code{\link{LD}}
}
