% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducing_list.R
\name{reducing_list}
\alias{reducing_list}
\title{Reducing haplotype or diplotype list elements}
\usage{
reducing_list(lst, nr_gene)
}
\arguments{
\item{lst}{A list. Data containing the possible diplotypes of each donor as a separate list element, can be obtained via \code{\link{all_options}}.}

\item{nr_gene}{An integer or a vector. Which genes \code{lst} needs to be reduced to.}
}
\value{
A list with the haplotypes or diplotypes of the specified genes only.
}
\description{
To reduce the hapltoype or diplotype lists to a specified set of genes.
}
