% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducing_mat.R
\name{reducing_mat}
\alias{reducing_mat}
\title{Reducing a diplotype probability matrix to the diplotypes of a single gene}
\usage{
reducing_mat(EM_mat, nr_gene, haplo = FALSE, noCNV = FALSE)
}
\arguments{
\item{EM_mat}{A matrix. The matrix updated via \code{\link{EM_algorithm}} for a combination of genes.}

\item{nr_gene}{An integer or a vector. To which genes \code{EM_mat} needs to be reduced.}

\item{haplo}{A logical scalar. Whether or not the matrix should be reduced to a haplotype probability matrix (\code{FALSE} is default).}

\item{noCNV}{A logical scalar, whether or not we want to split the CNVs into separate alleles (\code{FALSE} is default).}
}
\value{
A EM-algorithm matrix for the assigned gene(s).
}
\description{
Reduces a multi-locus updated diplotype probability matrix to a single-locus diplotype probability matrix. Support function of \code{\link{information_loss}}.
}
