% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_doubles.R
\name{remove_doubles}
\alias{remove_doubles}
\title{Removes redundant genotypic ambiguities}
\usage{
remove_doubles(dat)
}
\arguments{
\item{dat}{A vector. The genotyping data of a single gene.}
}
\value{
a vector with the unique genotypic ambiguities in the correct order.
}
\description{
Remove genotypic ambiguities are double and/or redundant and places the alleles of the remaining ambiguities in the correct order for further processing.
}
