% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_strange_symbols.R
\name{remove_strange_symbols}
\alias{remove_strange_symbols}
\title{Removal of specific strange symbols}
\usage{
remove_strange_symbols(dat)
}
\arguments{
\item{dat}{A vector. The genotype calls of a single gene.}
}
\value{
A vector where all the strange symbols are discarded from the genotypes.
}
\description{
Removing certain strange symbols (`*`, `?`, `:`) that originate in the genotype calls from altering the initial \emph{KIR} dataset.
}
\details{
Function is tailored to the occurrences in the obtained \emph{KIR} dataset. For extensions to other dataset always look at the patterns observed in that data and check if they match. \itemize{
  \item Some gene copies are preceded by a `*`.
  \item Some single gene genotype calls are followed by `:X?`.
  \item Some multi-gene genotype calls are followed by `:X?`.
}
}
