% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordering_grouping_results.R
\name{reordering_grouping_results}
\alias{reordering_grouping_results}
\title{Reordering gene names of reconstruction}
\usage{
reordering_grouping_results(
  names_lst,
  freqs_lst = NULL,
  freqs_mat = NULL,
  end_order
)
}
\arguments{
\item{names_lst}{A list. Haplotype notation (\emph{i.e., gene1-gene2-gene3}) of the order in which the genes are added into the reconstruction.}

\item{freqs_lst}{An optional list. Each list element is a separate vector of estimates. If \code{NULL}, than \code{freqs_mat} must be specified.}

\item{freqs_mat}{An optional matrix. Each column is a separate reconstruction, of which the column names need to be put in the common gene order. If \code{NULL}, than \code{freqs_lst} must be specified.}

\item{end_order}{A character. Haplotype notation of the order in which all reconstructions must be denoted.}
}
\value{

}
\description{
In order to compare the estimates from different reconstructions and/or starting genes, the haplotypes must be in the same order. Vector names will be put in the same common gene order.
}
