% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rounding}
\alias{rounding}
\title{Function for rounding of numbers}
\usage{
rounding(x, digits = 2, nsmall = 2)
}
\arguments{
\item{x}{A vector of numerical values that need to be rounded}

\item{digits}{An intiger. Indicating the number of decimal places}

\item{nsmall}{An intiger. The minimum number of digits to the right of the decimal point in formatting real/complex numbers in non-scientific formats}
}
\description{
Function for rounding of numbers
}
\details{
Values that are a 1000 or higher will be rounded scientifically with the number of digits of ndigits + 1
}
