% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_slashes.R
\name{splitting_slashes}
\alias{splitting_slashes}
\title{Splitting on the slashes ("/")}
\usage{
splitting_slashes(xx)
}
\arguments{
\item{xx}{A character. The genotype call of an individual.}
}
\value{
All possible genotypes compatible with the genotype call of the individual.
}
\description{
Support function of \code{\link{all_options}}. Function can only correctly handle genotype calls with allelic ambiguities (`/`) and cannot correctly handle arguments with a `+` or genotypic ambiguities (`|`).
}
