% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{SS}
\alias{SS}
\title{The first derivative of the likelihood for the information Loss}
\usage{
SS(HHOi, allele_freq)
}
\arguments{
\item{HHOi}{A vector. The values of an individual indicating how often the allele of interest occurs Homozygous-Heterozygous-Not.}

\item{allele_freq}{A numceric value. Allele frequencies of the allele of interest}
}
\value{
The product of the outcome of the first derivative for the individual of the allele of interest.
}
\description{
Calculating the first derivative of the likelihood for the information loss. Support function of \code{\link{information_loss}}.
}
