% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_maker.R
\name{table_maker}
\alias{table_maker}
\title{Make a table from multiple vectors}
\usage{
table_maker(
  ...,
  freqs_lst = NULL,
  freq_order = NULL,
  decreasing = TRUE,
  cnames = NULL,
  total = FALSE,
  rounding = 5
)
}
\arguments{
\item{...}{The vectors to be put next to each other based on their element names.}

\item{freqs_lst}{An optional list. Each list element contains a separate vector. If \code{NULL}, than vectors need to be supplied via \code{...}.}

\item{freq_order}{An optional vector. The names and order of the haplotypes to be plotted. If \code{NULL} all unique haplotypes will be plotted from high to low for the first vector.}

\item{decreasing}{An optional logical scalar. Whether or not the frequencies should be increasing or decreasing (default = \code{TRUE}). Only relevant when freq_order = \code{NULL}.}

\item{cnames}{An optional vector. Names of each of the vectors, must be supplied in the order in which the vectors are supplied.}

\item{total}{A logical scalar. Whether or not the last row contains a overview of how many elements each vector had.}

\item{rounding}{An integer. To how many digits all values need to be rounded.}
}
\value{
A table with each vector as a separate vector and each row a separate haplotype.
}
\description{
To visualize the effect of different methods on the estimates, the estimates need to be compared against each other. With \code{table_maker} sorts these vectors and pastes them next to each other.
}
