% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_2_table.R
\name{vector_2_table}
\alias{vector_2_table}
\title{Making a table from a vector}
\usage{
vector_2_table(vect)
}
\arguments{
\item{vect}{A vector. Two-gene haplotypes and their frequencies.}
}
\value{
A table with the alleles of the first gene as rows and the alleles of the second gene as columns.
}
\description{
Making from a two-gene haplotype vector a table with the genes separately as rows and columns.
}
