% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtContribMat.R
\name{dtContribMat}
\alias{dtContribMat}
\title{Gives the order of all haplotype combinations. Support function for \code{\link{Geno2EM_SB}}, \code{\link{EM_algorithm}} and for \code{\link{information_loss}}.}
\usage{
dtContribMat(Nhts)
}
\arguments{
\item{Nhts}{The number of unique haplotypes for the matrix}
}
\value{
An matrix with the order of haplotype combinations
}
\description{
Gives the order of all haplotype combinations. Support function for \code{\link{Geno2EM_SB}}, \code{\link{EM_algorithm}} and for \code{\link{information_loss}}.
}
