% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies_naive.R
\name{frequencies_naive}
\alias{frequencies_naive}
\title{Calculates frequencies for the naive analysis}
\usage{
frequencies_naive(lst, haplo = TRUE, adj_comb_vals = NULL, obs_diplos = TRUE)
}
\arguments{
\item{lst}{A list. Data containing the possible diplotypes of each donor as a separate list element, can be obtained via \code{\link{all_options}}.}

\item{haplo}{A logical scalar. Whether or not diplotype frequencies are estimated by first estimating haplotype frequenices, or immediately start by estimating diplotype frequencies (\code{TRUE} is default).}

\item{adj_comb_vals}{An optional vector. \lars{Currently unknown}.}

\item{obs_diplos}{A logical scalar. Whether or not diplotype frequencies are estimated for only the observed diplotypes, or also for all theoretically possible diplotypes (\code{TRUE} is default).}
}
\value{

}
\description{
To calculate haplotype and diplotype frequencies for the naive analysis.
}
