% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_haplotypes.R
\name{likelihood_haplotypes}
\alias{likelihood_haplotypes}
\title{Calculates the log-likelihood for the EM-algorithm}
\usage{
likelihood_haplotypes(mat, haplos_i, indicator = NULL, normalize = FALSE)
}
\arguments{
\item{mat}{A matrix. The probability vectors for each individual, can be obtained via \code{\link{EM-algorithm}}.}

\item{haplos_i}{A vector. The haplotype frequencies for the haplotypes in \code{mat}, can be obtained via \code{\link{EM-algorithm}}.}

\item{indicator}{An optinal matrix. How much each haplotype (row) contributes to a diplotype (columns). If \code{NULL}, an indicator matrix will be made.}

\item{normalize}{An logical scalar. \lars{Not sure what it does}.}
}
\value{
The (log-)likelihood of the EM-algorithm iteration and the indicator matrix made (or supplied as argument).
}
\description{
Calculates the log-likelihood based on the data analyzed with \code{\link{EM-algorithm}}.
}
\examples{

}
